% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{consensus_sequence}
\alias{consensus_sequence}
\title{Calculate consensus sequence of all original reads}
\usage{
consensus_sequence(i, j)
}
\arguments{
\item{i}{Index of sam-file}

\item{j}{Index of mapped virus}
}
\value{
None
}
\description{
In order to visualise the simulation of read sequences, this function produces a four-coloured graphic
in which the colours blue, green, red and yellow represent the nucleotides A, C, G and T.
The read sequences of the original sam-file are plotted in rows whereupon each column corresponds to a base position of the reference genome.
Below the mapped read sequences, the consensus sequence is displayed and at the bottom the subsection of the reference genome belonging to the reads.
The graphic is stored in the temporary directory.
}
\details{
Before using the function \code{consensus_sequence}, the functions \code{\link{create_directories_and_file_paths}} and \code{\link{extract_information_from_reference_genome}} must be executed.
}
\examples{
consensus_sequence ( i = 1 , j = 3 )
}
