% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{error_rates}
\alias{error_rates}
\title{Calculation of error rates}
\usage{
error_rates(file_indices)
}
\arguments{
\item{file_indices}{Indices of the mapped sam-files belonging to the simulated paired fastq-files}
}
\value{
None
}
\description{
In total, this function computes six different diagnostic parameters counting relative frequencies, two mapping rates and four error rates.
Furthermore, there are three absolute frequencies which count all mapped reads, reads that are mapped to an excel list of viruses and reads that are mapped correctly.
These values are calculated by using information of the mapping files of the simulated fastq-files from the previous step.
The old excel-file from the original mapping results is completed by six columns for the diagnostic parameters and saved as a new excel-file.
}
\examples{
error_rates ( file_indices = 1 )
}
