% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_information_from_reference_genome}
\alias{extract_information_from_reference_genome}
\title{Extract information from reference genome}
\usage{
extract_information_from_reference_genome(
  formatted.reference_genome_fasta.file_path
)
}
\arguments{
\item{formatted.reference_genome_fasta.file_path}{File path to formatted reference genome fasta file}
}
\value{
None
}
\description{
The function \code{extract_information_from_reference_genome}
needs the file path to the formatted reference genome fasta-file
(in which one sequence is stored in one line; in the original fasta-file
there could be a line break after 80 characters) and extracts information
about the NC numbers, species, sequences and sequence lengths.
These information are stored as global variables for the simulation procedure afterwards.
}
\examples{
extract_information_from_reference_genome(formatted.reference_genome_fasta.file_path = file.choose()) # Path of formatted reference genome fasta file
}
