% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mapping_bowtie2}
\alias{mapping_bowtie2}
\title{Mapping of fastq-files}
\usage{
mapping_bowtie2(
  file_indices = rep(NA_integer_, 0),
  reference_genome_index_bowtie2.directory,
  mapq_filter_threshold = 0,
  threads
)
}
\arguments{
\item{file_indices}{Indices of the simulated fast-file pairs that will be mapped.}

\item{reference_genome_index_bowtie2.directory}{# Prefix (folder) of bowtie2 index files}

\item{mapq_filter_threshold}{# Reads with a mapping quality below this threshold will be deleted from the sam-files.}

\item{threads}{Number of threads that are used for mapping}
}
\value{
None
}
\description{
The simulated fastq-files are used to be mapped against the artificial viral reference genome with bowtie2. After this step, one can filter out reads with a low mapping quality.
The new sam-file with filtered reads is stored separately in the same directory.
}
\examples{
library(rChoiceDialogs)
my.prefix = rchoose.dir() # Choose prefix (folder) of reference genome bowtie2 index files
my.prefix = gsub("\\\\\\\\","/",my.prefix)
my.prefix = paste0(my.prefix,"/")
mapping_bowtie2 ( file_indices = 1 , reference_genome_index_bowtie2.directory = my.prefix , mapq_filter_threshold = 2 , threads = 2 ) # Takes a few seconds to run!
}
