% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sim_fastq}
\alias{sim_fastq}
\title{Simulation of fastq-files}
\usage{
sim_fastq(
  file_indices = rep(NA_integer_, 0),
  read_counts,
  start_positions_and_read_lengths,
  seed
)
}
\arguments{
\item{file_indices}{Indices of the sam-files from which the rds-objects are used to simulate the paired fastq-files.}

\item{read_counts}{If the number of reads are "original", the number of reads generated per virus is identical to the read counts of the sam-file.
By specifying the parameter as "density", there is the possibility to change the number of reads generated randomly based on a kernel density estimation of the original number of reads mapped to the species.
In the latter case, the number of reads of the mapped species which are to be generated differs slightly from the original distribution.}

\item{start_positions_and_read_lengths}{If the start positions and read lengths are "original", the start positions and read lengths are taken from the original sam-files so that every read is used exactly once.
In contrary, if the parameter setting is "random", the reads are sampled with replacement so that some reads may be used multiple or zero times.}

\item{seed}{Seed of the random number generator used to create random base nucleotides and quality values based on the provided statistical distributions.}
}
\value{
None
}
\description{
This function produces paired fastq-files based on the statistical distributions of the original sam-files which are generated by the function \code{\link{sim_init}}.
The simulated paired fastq-files will be stored in the fastq-directory.
}
\details{
Do not use read_counts = "density" in combination with start_positions_and_read_lengths = "original"!
Before using the function \code{sim_fastq}, the functions \code{\link{create_directories_and_file_paths}}, \code{\link{extract_information_from_reference_genome}} and \code{\link{sim_init}} must be executed.
}
\examples{
\dontrun{sim_fastq ( file_indices = 1 , read_counts = "density", start_positions_and_read_lengths = "random", seed = 42 )} # Takes a few seconds to run.
}
