% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sim_init}
\alias{sim_init}
\title{Simulation initialisation}
\usage{
sim_init(file_indices = rep(NA_integer_, 0), type_of_sam_file)
}
\arguments{
\item{file_indices}{Indices of the sam-files to be analysed}

\item{type_of_sam_file}{"init" for original sam-file or "sim" for sam-file of simulated fastq-files.}
}
\value{
None
}
\description{
Before the actual simulation procedure can start, statistical distributions of the original mapping results must be calculated by the function \code{sim_init}.
The statistical distributions from the sam-files will be stored in the RDS objects directory, for each NGS run separately.
}
\details{
Before using the function \code{sim_init}, the functions \code{\link{create_directories_and_file_paths}} and \code{\link{extract_information_from_reference_genome}} must be executed.
}
\examples{
\dontrun{sim_init(file_indices=1,type_of_sam_file="init")} # Processing takes a few minutes of time!
\dontrun{sim_init(file_indices=1,type_of_sam_file="sim")} # Execute after mapping simulated fastq-files!

}
