% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sim_init_graphics}
\alias{sim_init_graphics}
\title{Graphics of simulation initialisation}
\usage{
sim_init_graphics(file_indices = rep(NA_integer_, 0))
}
\arguments{
\item{file_indices}{Indices of the sam-files to produce graphics of}
}
\value{
None
}
\description{
This function produces graphics based on the statistical distributions of the original sam-files which are generated by the function \code{\link{sim_init}}.
The function plots distributions of nucleotide, mapping quality, read length, start position and quality value of all mapped viruses together and top three mapped viruses.
and writes the plots into one pdf-file.
}
\details{
Before using the function \code{sim_init_graphics}, the functions \code{\link{create_directories_and_file_paths}}, \code{\link{extract_information_from_reference_genome}} and \code{\link{sim_init}} must be executed.
}
\examples{
sim_init_graphics ( file_indices = 1 )
}
