% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sim_sam_graphics}
\alias{sim_sam_graphics}
\title{Graphics of simulated sam-files}
\usage{
sim_sam_graphics(file_indices = rep(NA_integer_, 0))
}
\arguments{
\item{file_indices}{Indices of the sam-files to produce graphics of}
}
\value{
None
}
\description{
This function produces graphics based on the statistical distributions of the original sam-files and the sam-files produced by mapping of the simulated fastq-files
which are generated by the function \code{\link{sim_init}} with parameters "init" and "sim".
The function plots distributions of nucleotide, mapping quality, read length, start position and quality value of all mapped viruses together and top three mapped viruses
for both types of sam-files and writes the plots into one pdf-file.
}
\details{
Before using the function \code{sim_init_graphics}, the functions \code{\link{create_directories_and_file_paths}}, \code{\link{extract_information_from_reference_genome}} and \code{\link{sim_init}} with parameters "init" and "sim" must be executed.
}
\examples{
sim_sam_graphics ( file_indices = 1 )
}
